chore: csv에서 점수를 읽어들일 때 자료형을 정수에서 실수로 변경

+++ app.py
@@ -53,9 +53,9 @@ def add_csv():
     df = pd.read_csv(file, encoding="utf-8-sig")  # 업로드된 파일 바로 읽기
     for i in range(len(df)):
         sname = df.iloc[i]["이름"]
-        kor = int(df.iloc[i]["국어"])
-        eng = int(df.iloc[i]["영어"])
-        mat = int(df.iloc[i]["수학"])
+        kor = float(df.iloc[i]["국어"])
+        eng = float(df.iloc[i]["영어"])
+        mat = float(df.iloc[i]["수학"])
         db.insert_score(sname, kor, eng, mat)
     db.recalc_rank()
     return redirect(url_for("index"))
